Add isset check on the lastlogoff property
I've found that the lastlogoff property is missing if the account you're fetching information for is very new and hasn't been setup with a community profile. This change checks for the property using the built-in isset check. The value will fall back to the current server date/time if the property is missing from the api response.